ModelBit/requirements.txt
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0
modelbit==0.42.5